refactor: move contracts-ts tests from sdk to contracts
