+++ packages/contracts-ts/package.json
@@ -14,15 +14,27 @@
     "build": "tsup src/index.ts --format esm,cjs --dts",
     "dev": "tsup src/index.ts --format esm,cjs --watch --dts",
     "lint": "eslint src/**/*.ts",
-    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
+    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
+    "test": "jest",
+    "test:watch": "jest --watch"
   },
   "dependencies": {
-    "@cowprotocol/sdk-common": "workspace:*"
+    "@cowprotocol/sdk-common": "workspace:*",
+    "@cowprotocol/sdk-config": "workspace:*"
   },
   "devDependencies": {
+    "@cowprotocol/sdk-ethers-v5-adapter": "workspace:*",
+    "@cowprotocol/sdk-ethers-v6-adapter": "workspace:*",
+    "@cowprotocol/sdk-viem-adapter": "workspace:*",
+    "@types/jest": "^29.5.5",
     "@types/node": "^20.17.31",
     "eslint": "^8.48.0",
+    "ethers-v5": "npm:ethers@^5.7.2",
+    "ethers-v6": "npm:ethers@^6.13.7",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
     "tsup": "^7.2.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.2.2",
+    "viem": "^2.28.4"
   }
 }
