+++ packages/contracts-ts/tests/order.test.ts
@@ -1,5 +1,5 @@
 import { sepolia } from 'viem/chains'
-import { createAdapters, TEST_ADDRESS, TEST_PRIVATE_KEY, TEST_RPC_URL } from '../setup'
+import { createAdapters, TEST_ADDRESS, TEST_PRIVATE_KEY, TEST_RPC_URL } from './setup'
 import { ethers as ethersV5 } from 'ethers-v5'
 import * as ethersV6 from 'ethers-v6'
 import { setGlobalAdapter, TypedDataDomain } from '@cowprotocol/sdk-common'
@@ -17,7 +17,7 @@ import {
   encodeSigningScheme,
   decodeTradeFlags,
   encodeTradeFlags,
-} from '@cowprotocol/sdk-contracts-ts'
+} from '../src'
 import { privateKeyToAccount } from 'viem/accounts'
 import { createWalletClient, http } from 'viem'
