+++ packages/contracts-ts/tests/settlement.test.ts
@@ -1,5 +1,5 @@
 import { sepolia } from 'viem/chains'
-import { createAdapters, TEST_ADDRESS, TEST_PRIVATE_KEY, TEST_RPC_URL } from '../setup'
+import { createAdapters, TEST_ADDRESS, TEST_PRIVATE_KEY, TEST_RPC_URL } from './setup'
 import { ethers as ethersV5 } from 'ethers-v5'
 import * as ethersV6 from 'ethers-v6'
 import { setGlobalAdapter } from '@cowprotocol/sdk-common'
@@ -12,7 +12,7 @@ import {
   InteractionStage,
   Prices,
   TradeExecution,
-} from '@cowprotocol/sdk-contracts-ts'
+} from '../src'
 import { privateKeyToAccount } from 'viem/accounts'
 import { createWalletClient, http } from 'viem'
