+++ packages/sdk/package.json
@@ -19,13 +19,11 @@
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
   },
   "devDependencies": {
-    "@babel/preset-env": "7.27.2",
     "@cowprotocol/sdk-ethers-v5-adapter": "workspace:*",
     "@cowprotocol/sdk-ethers-v6-adapter": "workspace:*",
     "@cowprotocol/sdk-viem-adapter": "workspace:*",
     "@types/jest": "^29.5.5",
     "@types/node": "^20.17.31",
-    "babel-jest": "29.7.0",
     "ethers-v5": "npm:ethers@^5.7.2",
     "ethers-v6": "npm:ethers@^6.13.7",
     "jest": "^29.7.0",
